# basic-math-server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GlobalResponse_number_": {
				"properties": {
					"result": {
						"type": "number",
						"format": "double",
						"description": "{{description}}"
					},
					"error": {
						"type": "boolean",
						"description": "{{description}}",
						"example": false
					},
					"message": {
						"type": "string",
						"description": "{{description}}"
					}
				},
				"required": [
					"result",
					"error",
					"message"
				],
				"type": "object"
			},
			"AddBody": {
				"properties": {
					"numbers": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "{{description}}"
					}
				},
				"required": [
					"numbers"
				],
				"type": "object"
			},
			"DecreaseBody": {
				"properties": {
					"value": {
						"type": "number",
						"format": "double",
						"description": "{{description}}"
					},
					"from": {
						"type": "number",
						"format": "double",
						"description": "{{description}}"
					}
				},
				"required": [
					"value",
					"from"
				],
				"type": "object"
			},
			"MultiplyBody": {
				"properties": {
					"numbers": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "{{description}}"
					}
				},
				"required": [
					"numbers"
				],
				"type": "object"
			},
			"DivideBody": {
				"properties": {
					"by": {
						"type": "number",
						"format": "double",
						"description": "{{description}}"
					},
					"value": {
						"type": "number",
						"format": "double",
						"description": "{{description}}"
					}
				},
				"required": [
					"by",
					"value"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "basic-math-server",
		"version": "0.0.1",
		"description": "an api for doing basic math works",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "a.tanoomandian"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/basic/plus": {
			"post": {
				"operationId": "Add",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GlobalResponse_number_"
								}
							}
						}
					}
				},
				"description": "{{description}}",
				"tags": [
					"ریاضی"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "{{description}}",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddBody",
								"description": "{{description}}"
							}
						}
					}
				}
			}
		},
		"/basic/minus": {
			"post": {
				"operationId": "Decrease",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GlobalResponse_number_"
								}
							}
						}
					}
				},
				"description": "{{description}}",
				"tags": [
					"ریاضی"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "{{description}}",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DecreaseBody",
								"description": "{{description}}"
							}
						}
					}
				}
			}
		},
		"/basic/multiply": {
			"post": {
				"operationId": "Multiply",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GlobalResponse_number_"
								}
							}
						}
					}
				},
				"description": "{{description}}",
				"tags": [
					"ریاضی"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "{{description}}",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MultiplyBody",
								"description": "{{description}}"
							}
						}
					}
				}
			}
		},
		"/basic/divide": {
			"post": {
				"operationId": "Divide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GlobalResponse_number_"
								}
							}
						}
					}
				},
				"description": "{{description}}",
				"tags": [
					"ریاضی"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "{{description}}",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DivideBody",
								"description": "{{description}}"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}